+++ code/Core/Memory/Allocator.h
@@ -14,11 +14,11 @@ namespace Core
 		Allocator(void* pStart, size_t size);
 		virtual ~Allocator();
 
-		virtual void* Allocate(size_t size, uint8_t alignment) = 0;
+		virtual void* Allocate(size_t size, uint16_t alignment) = 0;
 		virtual void Free(void* pPtr) = 0;
 
 	protected:
-		uint8_t ComputeAlignmentOffset(const void* pPtr, uint8_t alignement) const;
+		uint16_t ComputeAlignmentOffset(const void* pPtr, uint16_t alignement) const;
 
 		char* m_pStart; //start pointer of the memory usable by this allocator
 		size_t m_size; //size of the memory usable by this allocator
